make current amplifier abstract
